+++ src/Pages/PageWithDetail.php
@@ -96,7 +96,7 @@ class PageWithDetail extends Page
         $filtersTitle = localize("filters.title");
         if (!($instance instanceof Author) && in_array('author', $filterLinks)) {
             $title = localize(phrase: "authors.title");
-            $href = $this->handler::page(Author::PAGE_ALL, $params);
+            $href = $this->handler::route(Author::ROUTE_ALL, $params);
             $relation = "authors";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Author::URL_PARAM] ??= 1;
@@ -104,7 +104,7 @@ class PageWithDetail extends Page
         }
         if (!($instance instanceof Language) && in_array('language', $filterLinks)) {
             $title = localize("languages.title");
-            $href = $this->handler::page(Language::PAGE_ALL, $params);
+            $href = $this->handler::route(Language::ROUTE_ALL, $params);
             $relation = "languages";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Language::URL_PARAM] ??= 1;
@@ -112,7 +112,7 @@ class PageWithDetail extends Page
         }
         if (!($instance instanceof Publisher) && in_array('publisher', $filterLinks)) {
             $title = localize("publishers.title");
-            $href = $this->handler::page(Publisher::PAGE_ALL, $params);
+            $href = $this->handler::route(Publisher::ROUTE_ALL, $params);
             $relation = "publishers";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Publisher::URL_PARAM] ??= 1;
@@ -120,7 +120,7 @@ class PageWithDetail extends Page
         }
         if (!($instance instanceof Rating) && in_array('rating', $filterLinks)) {
             $title = localize("ratings.title");
-            $href = $this->handler::page(Rating::PAGE_ALL, $params);
+            $href = $this->handler::route(Rating::ROUTE_ALL, $params);
             $relation = "ratings";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Rating::URL_PARAM] ??= 1;
@@ -128,7 +128,7 @@ class PageWithDetail extends Page
         }
         if (!($instance instanceof Serie) && in_array('series', $filterLinks)) {
             $title = localize("series.title");
-            $href = $this->handler::page(Serie::PAGE_ALL, $params);
+            $href = $this->handler::route(Serie::ROUTE_ALL, $params);
             $relation = "series";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Serie::URL_PARAM] ??= 1;
@@ -136,7 +136,7 @@ class PageWithDetail extends Page
         }
         if (in_array('tag', $filterLinks)) {
             $title = localize("tags.title");
-            $href = $this->handler::page(Tag::PAGE_ALL, $params);
+            $href = $this->handler::route(Tag::ROUTE_ALL, $params);
             $relation = "tags";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Tag::URL_PARAM] ??= 1;
@@ -148,7 +148,7 @@ class PageWithDetail extends Page
         }
         if (in_array('identifier', $filterLinks)) {
             $title = localize("identifiers.title");
-            $href = $this->handler::page(Identifier::PAGE_ALL, $params);
+            $href = $this->handler::route(Identifier::ROUTE_ALL, $params);
             $relation = "identifiers";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Identifier::URL_PARAM] ??= 1;
@@ -160,7 +160,7 @@ class PageWithDetail extends Page
         }
         if (in_array('format', $filterLinks)) {
             $title = localize("formats.title");
-            $href = $this->handler::page(Format::PAGE_ALL, $params);
+            $href = $this->handler::route(Format::ROUTE_ALL, $params);
             $relation = "formats";
             $this->addHeaderEntry($title, $filtersTitle, $href, $relation);
             $paging[Format::URL_PARAM] ??= 1;